requirements.txt
asgiref==3.8.1
sqlparse==0.5.0
typing_extensions==4.11.0
Django>=3.0,<4.0
psycopg2-binary==2.9.9
django-solo==2.2.0
pillow==10.3.0